Fixed 'file not found error in Xfoil function'

+++ BEM.py
@@ -102,9 +102,9 @@ class blade:
     def xfoil(self,inter = 200 , np = 220):
 
         # This function uses some Xfoils calculations to retrieve the values for Cl and Cd
-
         Coef = list()
-        os.remove("xfoil_output.txt")
+        if os.path.isfile("xfoil_output.txt"):
+            os.remove("xfoil_output.txt")
         itens  = [self.current_airfoil,self.current_alfa, self.current_alfa, '0', round(self.ma,2),int(self.re),inter,np]
 
         if platform.system() =='Linux':
@@ -126,7 +126,6 @@ class blade:
             os.system('del dump')
             os.system('del xfoil_input')
             os.system('del screen')
-            os.system('del xfoil_output')
             f=open('xfoil_input.txt','w')
             f.write('%s \n' % ('load'))
             f.write('%s \n' % (itens[0]))
@@ -164,28 +163,30 @@ class blade:
         xutr = list()
         xltr = list()
 
-        f      = open('xfoil_output.txt','r')
+        if os.path.isfile("xfoil_output.txt"):
+            f      = open('xfoil_output.txt','r')
 
-        for line in f:
-            if (i > 11):
+            for line in f:
+                if (i > 11):
 
-                aoa.append(line.strip().split()[0])
-                cl.append(line.strip().split()[1])
-                cd.append(line.strip().split()[2])
-                cdp.append(line.strip().split()[3])
-                cm.append(line.strip().split()[4])
-                xutr.append(line.strip().split()[5])
-                xltr.append(line.strip().split()[6])
-                #print((line.strip().split()[6]))
-                res = float(line.strip().split()[2])
-            else:
-                res = 1.0
-            i += 1
+                    aoa.append(line.strip().split()[0])
+                    cl.append(line.strip().split()[1])
+                    cd.append(line.strip().split()[2])
+                    cdp.append(line.strip().split()[3])
+                    cm.append(line.strip().split()[4])
+                    xutr.append(line.strip().split()[5])
+                    xltr.append(line.strip().split()[6])
+                    #print((line.strip().split()[6]))
+                    res = float(line.strip().split()[2])
+                else:
+                    res = 1.0
+                i += 1
 
-        f.close()
+            f.close()
 
             #colocar maximo
         j = 0
+
 
         if len(cl)==0:
             self.Cl =0.0000001
@@ -276,9 +277,9 @@ class blade:
             if self.correction :
 
                 if self.a[i] >self.a_critic :
-                    a.pop()
+                    self.a.pop()
                     K_h = 4*F*np.sin(Phi)**2/(sigma*Cn)
-                    a.append((1/2)*(2+K_h*(1-2*ac)-((K_h*(1-2*ac)+2)**2+4*(K_h*ac**2-1))**(1/2)))
+                    self.a.append((1/2)*(2+K_h*(1-2*self.a_critic)-((K_h*(1-2*self.a_critic)+2)**2+4*(K_h*self.a_critic**2-1))**(1/2)))
 
                 if self.a[i] <= (1/3):
                     Ct = 4*self.a[i]*(1-self.a[i])*F
